Fix -Wformat warnings on macos
Since macos uses unsigned long long instead of unsigned long for uint64_t,
the build log is flooded with lots of -Wformat warnings without this change.

+++ riscv/processor.cc
@@ -478,8 +478,8 @@ void processor_t::take_trap(trap_t& t, reg_t epc)
 #ifdef ENABLE_CHERI
     if (t.cause() == CAUSE_CHERI_TRAP) {
        cheri_t *cheri = (static_cast<cheri_t*>(get_extension()));
-       fprintf(stderr, "CHERI-DEBUG: PRV %lu = CCSR = 0x%lx\n", state.prv,
-       cheri->get_ccsr());
+       fprintf(stderr, "CHERI-DEBUG: PRV %lu = CCSR = 0x%lx\n", (long)state.prv,
+               (long)cheri->get_ccsr());
     }
 #endif
